third_party/blink/web_tests/platform/mac/external/wpt/html/semantics/forms/the-input-element/show-picker-disabled-readonly-expected.txt

Content-Type: text/plain
This is a testharness.js-based test.
[FAIL] input[type=button] showPicker() doesn't throw when readonly
  assert_false: User activation is consumed for non-readonly showPicker() call expected false got true
[FAIL] input[type=checkbox] showPicker() doesn't throw when readonly
  assert_false: User activation is consumed for non-readonly showPicker() call expected false got true
[FAIL] input[type=color] showPicker() doesn't throw when readonly
  assert_false: User activation is consumed for non-readonly showPicker() call expected false got true
[FAIL] input[type=file] showPicker() doesn't throw when readonly
  assert_false: User activation is consumed for non-readonly showPicker() call expected false got true
[FAIL] input[type=hidden] showPicker() doesn't throw when readonly
  assert_false: User activation is consumed for non-readonly showPicker() call expected false got true
[FAIL] input[type=image] showPicker() doesn't throw when readonly
  assert_false: User activation is consumed for non-readonly showPicker() call expected false got true
[FAIL] input[type=radio] showPicker() doesn't throw when readonly
  assert_false: User activation is consumed for non-readonly showPicker() call expected false got true
[FAIL] input[type=range] showPicker() doesn't throw when readonly
  assert_false: User activation is consumed for non-readonly showPicker() call expected false got true
[FAIL] input[type=reset] showPicker() doesn't throw when readonly
  assert_false: User activation is consumed for non-readonly showPicker() call expected false got true
[FAIL] input[type=submit] showPicker() doesn't throw when readonly
  assert_false: User activation is consumed for non-readonly showPicker() call expected false got true
Harness: the test ran to completion.

#EOF
#EOF
